refactor(plugin-betswirl): rename coinTossAction file
